Moved bot logic to 'bot' folder
