+++ src/service/main.js
@@ -1,11 +1,10 @@
-import Comm from "./comm";
-import settings from "./settings";
+import Comm from "./bot/comm";
+import settings from "./bot/settings";
 
 import storage from "storage";
 import db from "./db/db";
-import logger from "./db/logger";
 
-import robot from "./robot";
+import robot from "./bot/robot";
 
 import actions from "./actions";
 
@@ -81,55 +80,60 @@ Comm.listen("manager", function(action, data) {
 		Comm.sendMessage("backendError", {error});
 });
 
-for(var plugIndex = 0; plugs.enabledPlugs.length; plugIndex++) {
-	let plug = plugs.enabledPlugs[plugIndex],
-		plugInstantiator = await require(`./plugs/${plug}/${plug}`);
 
-	let settingsInterface = new settings(plug);
+async function bootstrap() {
+	for(var plugIndex = 0; plugs.enabledPlugs.length; plugIndex++) {
+		let plug = plugs.enabledPlugs[plugIndex],
+			plugInstantiator = await require(`./plugs/${plug}/${plug}`);
+	
+		let settingsInterface = new settings(plug);
+	
+		plugsInstances[plug] = {
+			settings: settingsInterface,
+			plug: new plugInstantiator(settingsInterface),
+			bot: false
+		}
+	}
 
-	plugsInstances[plug] = {
-		settings: settingsInterface,
-		plug: new plugInstantiator(settingsInterface),
-		bot: false
+	for (var i in plugs) {
+		var plugContainer = plugsInstances[i];
+		if (plugContainer.bot)
+			continue;
+
+		// Let's dance!
+		plugContainer.bot = new robot(plugContainer.settings, plugContainer.plug, i);
+		plugContainer.bot.start();
+
+		// Send errors to frontend when necessary
+		plugContainer.bot.addListener("error", (t, name, error) => {
+			Comm.sendMessage("backendError", {
+				plug: name,
+				data: error
+			});
+		});
+
+		// Reset errors on bot boot
+		plugContainer.bot.addListener("start", (t, name) => {
+			error = false;
+
+			Comm.sendMessage("backendError", {remove: true, plug: name});
+			// Update settings status
+			Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
+		});
+
+		// Send update to frontend. Ideally at this moment the user is showing a waiting screen during reboot
+		plugContainer.bot.addListener("reboot", (t, name) => {
+			Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
+		});
+
+		plugContainer.bot.addListener("stop", (t, name) => {
+			Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
+		});
 	}
 }
 
 // Start the bot when the browser is started!
-for (var i in plugs) {
-	var plugContainer = plugsInstances[i];
-	if (plugContainer.bot)
-		continue;
-
-	// Let's dance!
-	plugContainer.bot = new robot(plugContainer.settings, plugContainer.plug, i);
-	plugContainer.bot.start();
-
-	// Send errors to frontend when necessary
-	plugContainer.bot.addListener("error", (t, name, error) => {
-		Comm.sendMessage("backendError", {
-			plug: name,
-			data: error
-		});
-	});
-
-	// Reset errors on bot boot
-	plugContainer.bot.addListener("start", (t, name) => {
-		error = false;
-
-		Comm.sendMessage("backendError", {remove: true, plug: name});
-		// Update settings status
-		Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
-	});
-
-	// Send update to frontend. Ideally at this moment the user is showing a waiting screen during reboot
-	plugContainer.bot.addListener("reboot", (t, name) => {
-		Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
-	});
-
-	plugContainer.bot.addListener("stop", (t, name) => {
-		Comm.sendMessage("statusUpdate", {status: t.getStatus(), plug: name});
-	});
-}
+bootstrap()
 
 function getAllInitInfo() {
 	Comm.sendMessage("initInfo", {});
